reindex module
