+++ src/index.js
@@ -1,12 +1,8 @@
 export * as ast from 'ast';
 export * as builtin from 'builtin';
 export * as parser from 'parser';
+export * as module from 'module';
 export * as compiler from 'compiler';
 export * as analyzer from 'analyzer';
 export * as translator from 'translator';
 export * as plugin from 'plugin';
-
-import { CommonContext } from 'module';
-export var context = {
-    CommonContext
-};
