+++ src/module/index.js
@@ -1,0 +1,3 @@
+export { Resolver } from './Resolver';
+export * as context from './context';
+export * as loader from './loader';
